Move v0 add recipe models to one file
